+++ CHANGELOG.md
@@ -11,6 +11,7 @@ title: Changelog
   * removed the `options` parameter as it was not used anymore
   * add the possibility to provide a suffix instead of just being able to provide an event name =>
   It gives the possibility to create a suffix with a `tagName:suffix` directly if needed
+* Remove exports from package.json
 
 ## 6.8.0 (2021-06-01)